+++ src/missionMap/missionMap.test.ts
@@ -1,6 +1,6 @@
 import {TerrainTileMap} from "../hexMap/terrainTileMap";
 import {SquaddieID} from "../squaddie/id";
-import {HexCoordinate, HexGridTile, Integer} from "../hexMap/hexGrid";
+import {HexCoordinate, HexGridTile} from "../hexMap/hexGrid";
 import {HexGridMovementCost} from "../hexMap/hexGridMovementCost";
 import {SquaddieResource} from "../squaddie/resource";
 import {TraitCategory, TraitStatusStorage} from "../trait/traitStatusStorage";
@@ -13,9 +13,9 @@ describe('Mission Map', () => {
     beforeEach(() => {
         map = new TerrainTileMap({
             tiles: [
-                new HexGridTile(0 as Integer, -1 as Integer, HexGridMovementCost.singleMovement),
-                new HexGridTile(0 as Integer, 0 as Integer, HexGridMovementCost.singleMovement),
-                new HexGridTile(0 as Integer, 1 as Integer, HexGridMovementCost.doubleMovement),
+                new HexGridTile(0, -1, HexGridMovementCost.singleMovement),
+                new HexGridTile(0, 0, HexGridMovementCost.singleMovement),
+                new HexGridTile(0, 1, HexGridMovementCost.doubleMovement),
             ]
         });
 
@@ -35,7 +35,7 @@ describe('Mission Map', () => {
             terrainTileMap: map
         })
 
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 1 as Integer});
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 1});
 
         const squaddieMapCoordinate: HexCoordinate = missionMap.getSquaddieLocationById(torrinSquaddie.id);
         expect(squaddieMapCoordinate.q).toBe(0);
@@ -48,7 +48,7 @@ describe('Mission Map', () => {
         })
 
         let error: Error;
-        error = missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 1 as Integer});
+        error = missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 1});
         expect(error).toBeUndefined();
 
         const sirCamilSquaddie = new SquaddieID({
@@ -61,13 +61,13 @@ describe('Mission Map', () => {
             affiliation: SquaddieAffiliation.PLAYER,
         });
 
-        error = missionMap.addSquaddie(sirCamilSquaddie, {q: 0 as Integer, r: 1 as Integer});
+        error = missionMap.addSquaddie(sirCamilSquaddie, {q: 0, r: 1});
         expect(error.message.includes("already occupied")).toBeTruthy();
 
-        error = missionMap.addSquaddie(sirCamilSquaddie, {q: 2 as Integer, r: 1 as Integer});
+        error = missionMap.addSquaddie(sirCamilSquaddie, {q: 2, r: 1});
         expect(error.message.includes("not on map")).toBeTruthy();
 
-        error = missionMap.addSquaddie(sirCamilSquaddie, {q: 0 as Integer, r: -1 as Integer});
+        error = missionMap.addSquaddie(sirCamilSquaddie, {q: 0, r: -1});
         expect(error).toBeUndefined();
     });
 
@@ -76,21 +76,21 @@ describe('Mission Map', () => {
             terrainTileMap: map
         })
 
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 1 as Integer});
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 1});
 
-        let mapInformation = missionMap.getMapInformationForLocation({q: 0 as Integer, r: 1 as Integer});
+        let mapInformation = missionMap.getMapInformationForLocation({q: 0, r: 1});
         expect(mapInformation.q).toBe(0);
         expect(mapInformation.r).toBe(1);
         expect(mapInformation.squaddieId).toBe(torrinSquaddie.id);
         expect(mapInformation.tileTerrainType).toBe(HexGridMovementCost.doubleMovement);
 
-        mapInformation = missionMap.getMapInformationForLocation({q: 0 as Integer, r: -1 as Integer});
+        mapInformation = missionMap.getMapInformationForLocation({q: 0, r: -1});
         expect(mapInformation.q).toBe(0);
         expect(mapInformation.r).toBe(-1);
         expect(mapInformation.squaddieId).toBeUndefined();
         expect(mapInformation.tileTerrainType).toBe(HexGridMovementCost.singleMovement);
 
-        mapInformation = missionMap.getMapInformationForLocation({q: 0 as Integer, r: -5 as Integer});
+        mapInformation = missionMap.getMapInformationForLocation({q: 0, r: -5});
         expect(mapInformation.q).toBeUndefined();
         expect(mapInformation.r).toBeUndefined();
         expect(mapInformation.squaddieId).toBeUndefined();
@@ -102,7 +102,7 @@ describe('Mission Map', () => {
             terrainTileMap: map
         })
 
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 1 as Integer});
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 1});
 
         const squaddieMapCoordinate: HexCoordinate = missionMap.getSquaddieLocationById("id does not exist");
         expect(squaddieMapCoordinate.q).toBeUndefined();
@@ -111,32 +111,32 @@ describe('Mission Map', () => {
 
     it('can move a squaddie by updating its position', () => {
         const missionMap: MissionMap = new MissionMap({terrainTileMap: map});
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 0 as Integer})
-        missionMap.updateSquaddiePosition(torrinSquaddie.id, {q: 0 as Integer, r: 1 as Integer})
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 0})
+        missionMap.updateSquaddiePosition(torrinSquaddie.id, {q: 0, r: 1})
         const squaddieMapCoordinate: HexCoordinate = missionMap.getSquaddieLocationById(torrinSquaddie.id);
         expect(squaddieMapCoordinate.q).toBe(0);
         expect(squaddieMapCoordinate.r).toBe(1);
 
-        expect(missionMap.getSquaddieAtLocation({q: 0 as Integer, r: 0 as Integer})).toBeUndefined();
+        expect(missionMap.getSquaddieAtLocation({q: 0, r: 0})).toBeUndefined();
     });
 
     it('should be able to stay in place', () => {
         const missionMap: MissionMap = new MissionMap({terrainTileMap: map});
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 0 as Integer})
-        missionMap.updateSquaddiePosition(torrinSquaddie.id, {q: 0 as Integer, r: 0 as Integer})
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 0})
+        missionMap.updateSquaddiePosition(torrinSquaddie.id, {q: 0, r: 0})
         const squaddieMapCoordinate: HexCoordinate = missionMap.getSquaddieLocationById(torrinSquaddie.id);
         expect(squaddieMapCoordinate.q).toBe(0);
         expect(squaddieMapCoordinate.r).toBe(0);
 
-        expect(missionMap.getSquaddieAtLocation({q: 0 as Integer, r: 0 as Integer})).toBe(torrinSquaddie);
+        expect(missionMap.getSquaddieAtLocation({q: 0, r: 0})).toBe(torrinSquaddie);
     });
 
     it('should throw an error if wrong squaddie id', () => {
         const missionMap: MissionMap = new MissionMap({terrainTileMap: map});
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 0 as Integer})
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 0})
         const errorFound: Error | undefined = missionMap.updateSquaddiePosition("does not exist", {
-            q: 0 as Integer,
-            r: 1 as Integer
+            q: 0,
+            r: 1
         });
 
         expect(errorFound).toEqual(expect.any(Error));
@@ -145,11 +145,11 @@ describe('Mission Map', () => {
 
     it('should throw an error if off map', () => {
         const missionMap: MissionMap = new MissionMap({terrainTileMap: map});
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 0 as Integer})
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 0})
         const errorFound: Error | undefined = missionMap.updateSquaddiePosition(torrinSquaddie.id,
             {
-                q: 9001 as Integer,
-                r: -9001 as Integer
+                q: 9001,
+                r: -9001
             }
         );
 
@@ -169,9 +169,9 @@ describe('Mission Map', () => {
         });
 
         const missionMap: MissionMap = new MissionMap({terrainTileMap: map});
-        missionMap.addSquaddie(torrinSquaddie, {q: 0 as Integer, r: 0 as Integer})
-        missionMap.addSquaddie(sirCamilSquaddie, {q: 0 as Integer, r: 1 as Integer})
-        const errorFound = missionMap.updateSquaddiePosition(torrinSquaddie.id, {q: 0 as Integer, r: 1 as Integer})
+        missionMap.addSquaddie(torrinSquaddie, {q: 0, r: 0})
+        missionMap.addSquaddie(sirCamilSquaddie, {q: 0, r: 1})
+        const errorFound = missionMap.updateSquaddiePosition(torrinSquaddie.id, {q: 0, r: 1})
 
         expect(errorFound).toEqual(expect.any(Error));
         expect((errorFound as Error).message.includes(`occupied`)).toBeTruthy();
