+++ src/squaddie/turn.test.ts
@@ -1,4 +1,3 @@
-import {Integer} from "../hexMap/hexGrid";
 import {SquaddieActivity} from "./activity";
 import {ACTIVITY_PERFORM_FAILURE_REASON, SquaddieTurn} from "./turn";
 import {Trait, TraitCategory, TraitStatusStorage} from "../trait/traitStatusStorage";
@@ -12,13 +11,13 @@ describe('Squaddie turn and resources', () => {
             activitySpends2Actions = new SquaddieActivity({
                 id: "activitySpends2Actions",
                 name: "Power Attack",
-                actionsToSpend: 2 as Integer,
+                actionsToSpend: 2,
                 traits: new TraitStatusStorage({[Trait.ATTACK]: true}).filterCategory(TraitCategory.ACTIVITY)
             })
         })
 
         it('should start with 3 actions', () => {
-            expect(turn.getRemainingActions()).toBe(3 as Integer);
+            expect(turn.getRemainingActions()).toBe(3);
         });
         it('should spend 1 action by default', () => {
             turn.spendActionsOnActivity(
@@ -28,11 +27,11 @@ describe('Squaddie turn and resources', () => {
                     traits: new TraitStatusStorage({[Trait.ATTACK]: true}).filterCategory(TraitCategory.ACTIVITY)
                 })
             );
-            expect(turn.getRemainingActions()).toBe(2 as Integer);
+            expect(turn.getRemainingActions()).toBe(2);
         });
         it('should spend multiple actions if activity uses more', () => {
             turn.spendActionsOnActivity(activitySpends2Actions);
-            expect(turn.getRemainingActions()).toBe(1 as Integer);
+            expect(turn.getRemainingActions()).toBe(1);
         });
         it('should report when an activity cannot be spent', () => {
             turn.spendActionsOnActivity(activitySpends2Actions);
@@ -43,7 +42,7 @@ describe('Squaddie turn and resources', () => {
         it('should give 3 actions upon starting a new round', () => {
             turn.spendActionsOnActivity(activitySpends2Actions);
             turn.beginNewRound();
-            expect(turn.getRemainingActions()).toBe(3 as Integer);
+            expect(turn.getRemainingActions()).toBe(3);
         })
     });
 });
